+++ lib/kiq/pool/supervisor.ex
@@ -7,6 +7,8 @@ defmodule Kiq.Pool.Supervisor do
 
   @type options :: [config: Config.t(), name: GenServer.name()]
 
+  @resiliency_opts [sync_connect: true, exit_on_disconnection: true]
+
   @spec start_link(opts :: options()) :: Supervisor.on_start()
   def start_link(opts) do
     {name, opts} = Keyword.pop(opts, :name)
@@ -21,7 +23,11 @@ defmodule Kiq.Pool.Supervisor do
     children =
       for index <- 0..(size - 1) do
         name = Pool.worker_name(pool_name, index)
-        opts = Keyword.put(opts, :name, name)
+
+        opts =
+          opts
+          |> Keyword.put(:name, name)
+          |> Keyword.merge(@resiliency_opts)
 
         Supervisor.child_spec({Redix, {host, opts}}, id: name)
       end