+++ test/kiq/pool_test.exs
@@ -15,7 +15,7 @@ defmodule Kiq.PoolTest do
 
   describe "checkout/1" do
     test "a random connection pid is returned" do
-      config = %Config{pool_size: 2}
+      config = %Config{pool_size: 2, pool_name: PoolTest.Pool}
       name_a = Pool.worker_name(config.pool_name, 0)
       name_b = Pool.worker_name(config.pool_name, 1)