+++ test/kiq/reporter/supervisor_test.exs
@@ -1,12 +1,12 @@
 defmodule Kiq.Reporter.SupervisorTest do
   use Kiq.Case, async: true
 
-  alias Kiq.{Config, EchoClient, EchoConsumer}
+  alias Kiq.{EchoClient, EchoConsumer}
   alias Kiq.Reporter.Supervisor, as: ReporterSupervisor
 
   describe "start_link/1" do
     test "producer and consumer children are managed for the queue" do
-      config = Config.new(client_name: Echo, reporter_name: Kiq.Rep)
+      config = config(client_name: Echo, reporter_name: Kiq.Rep)
 
       {:ok, _ci} = start_supervised({EchoClient, name: Echo})
       {:ok, sup} = start_supervised({ReporterSupervisor, config: config})
@@ -24,7 +24,7 @@ defmodule Kiq.Reporter.SupervisorTest do
     end
 
     test "extra reporters are also supervised" do
-      config = Config.new(reporter_name: Kiq.Rep, reporters: [], extra_reporters: [EchoConsumer])
+      config = config(reporter_name: Kiq.Rep, reporters: [], extra_reporters: [EchoConsumer])
 
       {:ok, sup} = start_supervised({ReporterSupervisor, config: config})